rename helperFunctions to util
